test(realworld): fix dependencies & move setNav to app
