+++ test/realworld/src/app.html
@@ -1,7 +1,3 @@
-<import from="./shared/layouts/header-layout"></import>
-<import from="./shared/layouts/footer-layout"></import>
-<import from="./components/home/home"></import>
-
-<header-layout router.bind="router"></header-layout>
+<header-layout></header-layout>
 <au-viewport name="main" default="home"></au-viewport>
 <footer-layout></footer-layout>
