+++ test/realworld/src/resources/elements/article-preview.html
@@ -1,6 +1,4 @@
 <import from="../../shared/buttons/favorite-button"></import>
-<import from="../../components/profile/profile"></import>
-<import from="../../components/article/article"></import>
 
 <div class="article-preview">
   <div class="article-meta">
